+++ paper/paper.bib
@@ -1,0 +1,61 @@
+@book{Hernan:2020,
+    Author = {{Hernan}, M.A. and {Robins}, J.M. (2020).},
+    Title = {{Causal Inference: What If}},
+    Year = 2020,
+    Publisher = {Chapman and Hall/CRC},
+    City = {Boca Raton}
+}
+
+@book{Pearl:2009,
+    Author = {Pearl, J.},
+    Title = {Causality: Models, Reasoning and Inference},
+    Publisher = {Cambridge University Press},
+    Year = 2009,
+    City = {Cambridge}
+}
+
+@book{Bareinboim:2020,
+    Author = {{Bareinboim}, E. and {Correa}, J.D. and {Ibeling}, D. and {Icard}, T.},
+    Title = {{On Pearl`s Hierarchy and the Foundations of Causal Inference}},
+    Year = 2020,
+    Booktitle = {{Probabilistic and Causal Inference: the Works of Judea Pearl}},
+    Publisher = {{ACM Books}}
+}
+
+@book{Koller:2009,
+    Author = {{Koller}, D. and {Friedman}, N.},
+    Title = {{Probabilistic Graphical Models: Principles and Techniques}},
+    Year = 2009,
+    Publisher = {{MIT Press}},
+    City = {Cambridge}
+}
+
+@inproceedings{Richardson:2012,
+    Title = {Nested Markov Properties for Acyclic Directed Mixed Graphs},
+    Author = {{Richardson}, T.S. and {Evans}, R.J. and {Robins}, J.M. and {Shpitser}, I.},
+    Booktitle = {{UAI'12: Proceedings of the Twenty-Eighth Conference on Uncertainty in Artificial Intelligence}},
+    Publisher = {{ACM}},
+    Year = 2012,
+    Pages = 13
+}
+
+@article{Little:2019,
+    Title = {{Causal Bootstrapping}},
+    Author = {{Little}, M.A. and {Badawy}, R.},
+    Year = 2019,
+    Pages = {arXiv:1910.09648}
+}
+
+@article{Sturchio:2020,
+    Title = {{Phenotype-Agnostic Molecular Subtyping of Neurodegenerative Disorders: The Cincinnati Cohort Biomarker Program (CCBP)}},
+    Year = 2020,
+    Author = {{Sturchio}, A. and {Espay}, A.J.},
+    Journal = {{Frontiers in Aging Neuroscience}},
+    Volume = 12,
+    Pages = 324
+}
+
+
+
+
+
