### 2022-04-01 #4286 @max-little
Hi Kyle
Sorry about that, paper files have now been uploaded, you can find them in
the /paper subdirectory of the repository. Is that OK?
Best
Max
