+++ README.md
@@ -1,3 +1,5 @@
+![GRAPL logo](https://raw.githubusercontent.com/max-little/GRAPL/master/grapl.png)
+
 # GRAPL
 ### GRAPL: A computational library for nonparametric structural causal modelling, analysis and inference
 
