### 2022-04-02 #4286 @kyleniemeyer
@max-little thanks, the paper does build ok. However, could you look for missing DOIs in your references?
