+++ paper/paper.bib
@@ -11,12 +11,14 @@
     Title = {Causality: Models, Reasoning and Inference},
     Publisher = {Cambridge University Press},
     Year = 2009,
+	doi = {10.1017/S0266466603004109},
     City = {Cambridge}
 }
 
 @book{Bareinboim:2020,
     Author = {{Bareinboim}, E. and {Correa}, J.D. and {Ibeling}, D. and {Icard}, T.},
     Title = {{On Pearl`s Hierarchy and the Foundations of Causal Inference}},
+	doi = {10.1145/3501714.3501743},
     Year = 2020,
     Booktitle = {{Probabilistic and Causal Inference: the Works of Judea Pearl}},
     Publisher = {{ACM Books}}
@@ -52,10 +54,17 @@
     Author = {{Sturchio}, A. and {Espay}, A.J.},
     Journal = {{Frontiers in Aging Neuroscience}},
     Volume = 12,
-    Pages = 324
+    Pages = 324,
+	doi = {10.3389/fnagi.2020.553635}
 }
 
+@article{Shpitser:2008,
+	Title = {Complete Identification Methods for the Causal Hierarchy},
+	Year = 2008,
+	Journal = {Journal of Machine Learning Research},
+	Pages = 1941-1979
+}
 
-
+	
 
 
