### 2022-04-02 #4286 @max-little
Hi Kyle
Revised versions of the paper and bib files now uploaded. Hope these
are suitable.
Best
Max
