+++ paper/paper.bib
@@ -62,9 +62,6 @@
 	Title = {Complete Identification Methods for the Causal Hierarchy},
 	Year = 2008,
 	Journal = {Journal of Machine Learning Research},
-	Pages = 1941-1979
+	Pages = "1941-1979"
 }
-
-	
-
 
