### 2022-05-08 #4286 @max-little
Hi Nikoleta

Many thanks for taking the time to review the package. Of course, we are
happy to do our best to respond to any constructive criticism as best we
can.

Kind regards
Max

On Sun, 8 May 2022 at 19:02, Nikoleta Glynatsi ***@***.***>
wrote:

> Hey @max-little <https://github.com/max-little> 👋🏻
>
> I took some time to review your submission, and I believe that your
> package/repository requires a lot of work to get it in good shape for
> publication. After consulting with one of our associate editors we believe
> that it would be better to put the publication on hold for now and to
> resume once required changes have been made.
>
> I am going to list all the required changes in my review (will comment on
> this issue). What do you think about this? Are you okay if we proceed? I
> will post my review as soon as possible.
>
> @Athene-ai <https://github.com/Athene-ai> this also means that I would
> have to remove you as a reviewer and invite you again in the future. I hope
> that’s okay with you. I am sorry for the inconvenience.
>
