### 2022-05-15 #4286 @Nikoleta-v3
Apologises for the delay (had a very busy week 😪) nevertheless here is my review,

### Installation

At the moment the only way to use the `grapl` is to download the repository and have my scripts in the same folder as `grapl` (correct me if I am wrong). You need to package the project such that it can be installable and it can be used within a Python environment regardless of the location of the script, notebook, etc. There are several ways to do this.

- This is a more generic blog post: https://betterscientificsoftware.github.io/python-for-hpc/tutorials/python-pypi-packaging/
- This is the official Python Packaging Projects https://packaging.python.org/en/latest/tutorials/packaging-projects/

Once you have done that you will need to have proper installation notes. Not just "download the folder", but also cloning the project, installing the project and how to import it.

### Requirements

The requirement of your package is `ply`. As far as I can see `ply` is pip installable. I am not sure then why the source files of this package are part of your repository.

As part of packaging python code (links above) you can also specify the requirements of a package (creating a `requirements.txt` file). There we specify the version of the package, and when a user instals your package the requirement will be installed as well.

### Tests

Your package currently has some tests in the file `unit_tests.py` you need to document in the `readme` how to run the test suite.

Moreover, (this optional is just a recommendation) you should check out the package Pytest https://docs.pytest.org/en/7.1.x/. It will give you a better output and more information when you run your tests. Because of they way you have written your tests it should be easy to move to pytests.

For example:

```python
 grapl_obj = dsl.GraplDSL()

 grapl_filename = 'cyclic_graph.grapl'
 G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
 assert not G.isdag()
 assert not G.isayclic()
```

will become:

```python
def test_GraplDSL():
   grapl_obj = dsl.GraplDSL()

   grapl_filename = 'cyclic_graph.grapl'
   G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
   assert not G.isdag()
   assert not G.isayclic()
```

and from the terminal you run the tests with the command `pytest unit_tests.py` instead of `python unit_tests.py`.

## Documentation

I realise that the tutorial files are there to serve as a “Documentation” but they are not too self explanatory. You need to add some explanation about the methods you introduce in the tutorials and their outputs.

Ideally your package would have an online documentation, however, that is not a hard condition. If you would like to keep the tutorials as `py` files that’s okay but if you moved to markdown or jupyter notebooks it will be easier to include more text, equations (if needed), diagrams etc.

The `tutorial.py` files bring me to my next point:

### Structure

There are currently way too many not important files in the root of the direction. All the `grapl` files and `tutorial.py` should be in folders. For example `static` and `docs` respectively. The root should be cleaner.

### Additional files

A contribution file is necessary. Contributions files are there to encourage people to use your package and become part of your community. Contribution files don’t have to be too complex. Here is an example: https://github.com/OttoStruve/muler/blob/main/CONTRIBUTING.md. This is a project that was recently published with us.

Note that in your contributions files you will need to point to the sections 'Installation ' and `Running tests`.

### README

The release notes are of course fine but as this list keeps getting bigger it might be something you want to consider removing.

Another minor comment, I am using the night mode on Github and your logo is not visible.
