### 2022-05-15 #4286 @Nikoleta-v3
@max-little I recommend that you go over these points and let me know if something is not clear. If you are happy to apply all these changes then I (or you) should open issues in your repository for each respective point. This way it will be easier for me to check your changes 👍🏻

Once all the above have been addressed then we can continue 😄
