### 2022-05-28 cc9dd60
Delete grapl directory
