+++ grapl/algorithms.py
@@ -1,277 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Implements algorithms for processing of structural causal models represented
-by acyclic directed mixed graphs (ADMGs).
-
-(CC BY-SA 4.0) 2021. If you use this code, please cite: M.A. Little, R. Badawy,
-2019, "Causal bootstrapping", arXiv:1910.09648
-"""
-
-import grapl.admg as admg
-import grapl.expr as expr
-import grapl.util as util
-import copy as cp
-
-
-def idfixing(G, X={}, Y={}):
-    """Implementation of Richardson et al.'s Theorem 60 for causal effect identification
-       from given ADMG. See: T.S. Richardson, J.M. Robins, I. Shpitser, 2012:
-       "Nested Markov properties for acyclic directed mixed graphs", UAI'12
-    """
-
-    # Must supply the set of interventional (cause) variables
-    if not X:
-        return '', None, False
-
-    # Collate variable names
-    V = G.nodes()
-
-    # Handle effects
-    if not Y:
-        Y = V.difference(X)
-
-    # Find Y* = an_G(V)_{V\A}(Y)
-    A = X
-    VdA = V.difference(A)
-    G_sub_VdA = G.sub(VdA)
-    Y_star = G_sub_VdA.an(Y)
-
-    # Find subgraph G(V)_Y* and its districts
-    G_sub_Ys = G.sub(Y_star)
-    D_y = G_sub_Ys.districts()
-
-    # For each district, determine post-fixed ADMG and corresponding expression
-    D_exprs = []
-    for D in D_y:
-
-        # Nodes to fix for this district, fixing transform: phi_{V\D}(p(V;G_V))
-        VdD = V.difference(D)
-
-        # Initialize accumulated ADMG, G(V), and district expression, p(V)
-        G_fixed = cp.deepcopy(G)
-        D_expr = expr.Expr(num=[V])
-
-        # Try to apply fixing sequence
-        nodes_fix = VdD.copy()
-        while (len(nodes_fix) > 0):
-            
-            # Find a valid fixing node in sequence
-            fixable_nodes = {node for node in nodes_fix if G_fixed.isfixable(node)}
-            node_valid = (len(fixable_nodes) > 0)
-            if node_valid:
-                fix = fixable_nodes.pop()
-            else:
-                break
-            
-            # Fix or marginalize (Corollary 33) the accumulated expression
-            dispa_fix = G_fixed.dispa(fix)
-            fix_ch = G_fixed.ch({fix})
-            if len(fix_ch) == 0:
-                D_expr.fixmarginal(fix)
-            else:
-                D_expr.fix(fix, dispa_fix)
-
-            # Fix node (remove all its incoming edges) in accumulated ADMG,
-            # remove from remaining fixing sequence
-            G_fixed.fix(fix)
-            nodes_fix = nodes_fix.difference({fix})
-
-        if not node_valid:
-            break
-        
-        D_expr.simplify()
-        D_exprs.append(D_expr)
-
-    # At this point, if no valid fixing node, intervention is not identifiable
-    if not node_valid:
-        return '', None, False
-    else:
-        Y_mrg = Y_star.difference(Y)
-        id_expr = expr.Expr(mrg=Y_mrg)
-        for ex in D_exprs:
-            clash_vars = id_expr.mrg.difference(id_expr.mrg.difference(ex.mrg))
-            for old_var in clash_vars:
-                new_var = old_var + chr(39)
-                ex.subsvar(old_var, new_var)
-            id_expr.num = id_expr.num + ex.num
-            id_expr.den = id_expr.den + ex.den
[…]
