### 2022-05-28 ab48099
Delete tutorial2.py
