### 2022-05-28 4df5c96
Delete tutorial3.py
