### 2022-05-28 ada0842
Add files via upload
Major restructuring in order to fit within standard Python package guidelines.
