+++ grapl/admg.py
@@ -1,0 +1,292 @@
+# -*- coding: utf-8 -*-
+"""
+A small library of functions for representing, analyzing and processing
+acyclic directed mixed graphs for structural causal modelling.
+
+(CC BY-SA 4.0) 2021. If you use this code, please cite: M.A. Little, R. Badawy,
+2019, "Causal bootstrapping", arXiv:1910.09648
+"""
+
+from queue import Empty
+import grapl.util as util
+
+class ADMGNode():
+    """An ADMG node (random variable) object."""
+    def __init__(self, parents={}, bidirects={}, children={}, properties={}):
+        self.parents = set(parents)
+        self.children = set(children)
+        self.bidirects = set(bidirects)
+        self.properties = set(properties)
+
+class ADMG():
+    """An ADMG graph object."""
+    def __init__(self, title=''):
+        self.title = title
+        self.vars = dict()
+
+    def settitle(self, title=''):
+        """Set graph title."""
+        self.title = title
+
+    def addvar(self, name='', parents={}, bidirects={}, children={}, properties={}):
+        """Add a variable to a graph with optional parents, bidirects and/or children."""
+        if name != '':
+            self.vars[name] = ADMGNode(parents,bidirects,children,properties)
+        else:
+            raise ValueError('admg:addvar: Invalid node name')
+
+    def addedges(self, name='', parents={}, bidirects={}, children={}):
+        """Add parents, bidirects and/or child edges to a graph."""
+        if name != '':
+            self.vars[name].parents = self.vars[name].parents.union(parents)
+            self.vars[name].children = self.vars[name].children.union(children)
+            self.vars[name].bidirects = self.vars[name].bidirects.union(bidirects)
+        else:
+            raise ValueError('admg:addedges: Invalid node name')
+
+    def connect(self):
+        """Ensure all parent/child/bidirect edges are consistent across the graph."""
+        for node in set(self.vars):
+            for p in self.pa({node}):
+                self.vars[p].children = self.ch({p}).union({node})
+            for b in self.bi({node}):
+                self.vars[b].bidirects = self.bi({b}).union({node})
+
+    def display(self):
+        """Pretty-print a text representation of the graph."""
+        nodes = set(self.vars)
+        print('Title: ' + self.title)
+        print('Vars: ' + util.csepstr(nodes))
+        print('Parents:')
+        for node in nodes:
+            parents = self.pa({node})
+            if len(parents) > 0:
+                print(node,'<-', util.csepstr(parents))
+        print('Children:')
+        for node in nodes:
+            children = self.ch({node})
+            if len(children) > 0:
+                print(node,'->', util.csepstr(children))
+        print('Bidirects:')
+        for node in nodes:
+            bidirects = self.bi({node})
+            if len(bidirects) > 0:
+                print(node,'<-->', util.csepstr(bidirects))
+
+    def nodes(self):
+        """Return set of all nodes (variables)."""
+        return set(self.vars)
+
+    def sub(self, subvars):
+        """Find a subgraph of a graph."""
+        sub = ADMG(self.title)
+        for node in set(self.vars):
+            if node in subvars:
+                parents = self.pa({node}).intersection(subvars)
+                bidirects = self.bi({node}).intersection(subvars)
+                sub.vars[node] = ADMGNode(parents,bidirects,self.vars[node].properties)
+        sub.connect()
+        return sub
+    
+    def ch(self, nodes):
+        """Find the union of all children of a set of nodes."""
+        try:
+            children = set()
+            for node in nodes:
+                children = children.union(self.vars[node].children)
+            return children
+        except KeyError as err:
+            raise ValueError(f"admg.ch:Var '{node}' does not exist")
+
+    def pa(self, nodes):
[…]

[…]
