### 2022-05-28 #4286 @max-little
Hi Nikoleta

Thank you for all your comments, which were very helpful. I have been busy attending to all of them in detail; you can now find a completely restructured and reworked version of the library ready on Github. As requested, I've opened issues for each item listed in your comments above, hopefully these are all satisfactory now. Of course, do let me know if there's any further clarifications required.

Once again, thank you so much for your time and effort, this is very much appreciated.

Best
Max
