+++ grapl/dsl.py
@@ -1,0 +1,139 @@
+# -*- coding: utf-8 -*-
+"""
+Implements a lexer and parser of a simple domain-specific language
+(GRAPL -- GRAph Processing Language) for describing acyclic directed mixed graphs.
+A quasi-EBNF description of the language goes as follows, where 'grapl' is the top
+level rule:
+
+        grapl = TITLE EOC graph | graph
+        graph = graph command | command
+        command = NODE EOC | NODE DIR_EDGE NODE EOC | NODE BI_EDGE NODE EOC
+
+The terminal identifiers are:
+
+        TITLE: any string of characters enclosed in quotes
+        NODE: any non-empty string of alphabetical characters including underline,
+              followed optionally by any natural number
+        EOC: ';', the end of command character
+        DIR_EDGE: '->', indicating a directed edge
+        BI_EDGE: '<->', indicating a bidirected edge
+
+(CC BY-SA 4.0) 2021. If you use this code, please cite: M.A. Little, R. Badawy,
+2019, "Causal bootstrapping", arXiv:1910.09648
+"""
+
+import grapl.ply.lex as lex
+import grapl.ply.yacc as yacc
+import grapl.admg as admg
+import copy
+
+
+class GraplDSL():
+    """A GRAPL lexer and parser object, implemented using PLY."""
+    def __init__(self):
+        self.lexer = lex.lex(module=self)
+        self.parser = yacc.yacc(module=self)
+
+    tokens = (
+        'NODE',
+        'DIR_EDGE',
+        'BI_EDGE',
+        'EOC',
+        'TITLE'
+    )
+
+    t_DIR_EDGE = '->'
+    t_BI_EDGE = '<->'
+    t_EOC = ';'
+    t_ignore = ' \t'
+
+    def t_TITLE(self,t):
+        r'\"(.)*\"'
+        return t
+
+    def t_NODE(self,t):
+        r'([a-zA-Z_])+(\d)*'
+        return t
+
+    def t_newline(self,t):
+        r'\n+'
+        t.lexer.lineno += len(t.value)
+
+    def t_error(self,t):
+        raise SyntaxError("Grapl lexical error at line: " + str(t.lineno) + ", near: " + t.value)
+
+    def p_grapl(self,p):
+        '''grapl : TITLE EOC graph
+                 | graph'''
+        if len(p) == 4:
+            self.admg.settitle(p[1])
+
+    def p_graph(self,p):
+        '''graph : graph command
+                | command'''
+        pass
+
+    def p_command(self,p):
+        '''command : NODE EOC
+                | NODE DIR_EDGE NODE EOC
+                | NODE BI_EDGE NODE EOC'''
+        if len(p) == 3:
+            self.admg.addvar(p[1])
+        elif p[2] == self.t_DIR_EDGE:
+            if p[1] not in self.admg.vars:
+                print("Grapl error: node " + p[1] + " not already defined at line: " + str(p.lineno(1)))
+            elif p[3] not in self.admg.vars:
+                print("Grapl error: node " + p[3] + " not already defined at line: " + str(p.lineno(3)))
+            else:
+                self.admg.addedges(p[3],parents={p[1]})
+        elif p[2] == self.t_BI_EDGE:
+            if p[1] not in self.admg.vars:
+                print("Grapl error: node " + p[1] + " not already defined at line: " + str(p.lineno(1)))
+            elif p[3] not in self.admg.vars:
+                print("Grapl error: node " + p[3] + " not already defined at line: " + str(p.lineno(3)))
+            else:
+                self.admg.addedges(p[3],bidirects={p[1]})
+        
+    def p_error(self,p):
+        raise SyntaxError("Grapl syntax error at line: " + str(p.lineno) + ", near: " + p.value)
+
+    def readgrapl(self, scm_string=''):
[…]
