### 2022-06-29 #4286 @adavidzh
> [...] willing to review this submission for JOSS?

Thanks for the invitation. I'm afraid that at this time I am not able to.
