### 2022-07-01 #4286 @Nikoleta-v3
👋 @jhwohlgemuth & @fAndreuzzi would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "GRAPL: A computational library for nonparametric structural causal modelling, analysis and inference". You can find more information at the top of this Github issue 🆙
