### 2022-07-01 #4534 @editorialbot
[REVIEW]: GRAPL: A computational library for nonparametric structural causal modelling, analysis and inference
**Submitting author:** @max-little (<a href="http://orcid.org/0000-0002-1507-3822">Max A. Little</a>)
**Repository:** https://github.com/max-little/GRAPL
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.0
