### 2022-07-01 #4534 @Nikoleta-v3
👋🏼 @max-little  @Athene-ai, @fAndreuzzi this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

These checklists contain the JOSS requirements ✅  As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines.
