### 2022-07-02 2cd8640
rephrasing README. introduce citation file
