+++ CITATION.cff
@@ -1,0 +1,14 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: "Little"
+  given-names: "Max A."
+  orcid: "0000-0002-1507-3822"
+- family-names: "Badawy"
+  given-names: "Reham"
+  orcid: ""
+title: "Causal bootstrapping"
+version: 0.0.1
+doi: 10.48550/arXiv.1910.09648
+date-released: 2019-10-21
+url: "https://arxiv.org/abs/1910.09648"
