### 2022-07-02 max-little/GRAPL#10 @fandreuz
Use `pytest` or similar tools for the test suite
I would suggest to use a testing tool instead of what you're doing now in order to get better test results and maintainability. Anyone would do the work, [pytest](https://docs.pytest.org/en/7.1.x/), unittest, ...

Let me know if you need help with the setup @max-little, I would suggest pytest since it's the kind-of standard now in open source.

Part of openjournals/joss-reviews#4534
