### 2022-07-02 max-little/GRAPL#10 @fandreuz
Also I would suggest using the standard directory architecture for Python projects, i.e.

```
(root)
--- grapl/
--- tests/
      --- test_data/
      --- test1.py
      --- test2.py
--- tutorials/
```

In order to clean the source folder.
