### 2022-07-02 #4534 @fAndreuzzi
Hi @max-little, I'm going to help with the review for this paper. I've already submitted some suggestions to your repository such that it's easier for you to track what needs to be changed. For now it's only code structure/tooling, I'll review the code later.

I see that your code is not on PyPI, do you plan to upload it there such that your users can install GRAPL via `pip install grapl`?
