### 2022-07-02 max-little/GRAPL#10 @fandreuz
Replying to the comment by @max-little under #9:

> Re: PyTest, we have explored this and from long experience, want to avoid
going down the Python package interdependency route where possible. This is
one downside of OS development e.g. lack of central control over package
quality, and the resulting subtle and not so subtle bugs which arise as a
result. Our preference is to ensure our code is as self-contained as
possible. Of course, that's not always feasible.

Okay, but I would strongly suggest to have testing code and testing data separated from the actual source code for the package, which is what I suggest here: https://github.com/max-little/GRAPL/issues/10#issuecomment-1172864084
