### 2022-07-02 max-little/GRAPL#10 @max-little
Hi Francesco

Many thanks for the feedback. As part of the design process, we spent
significant effort crafting a folder structure which we believe
accurately conveys the conceptual organization of the software. The
idea is that, conforming to literate coding principles, as far as is
possible, the code itself (along with what is included and what is
excluded), is "self-describing". The unit tests provided here, are
meant to be included as part of the package itself, because they are
an integral part of the modular structure of the library, and serve an
important pedagogical purpose. Similar arguments hold for the
tutorials, sharing test data (graphs) with the unit tests.

I can certainly understand the motivation for excluding tests from a
package, but we would argue that this would sever the important link
between the tests and the functionality of the library. The scenario
we have often encountered in writing research software like this, is
that users tweak/adapt the code locally for their own purposes, and in
the process tend to break it (naturally, as they are experimenting).
Running the tests is a very quick way of diagnosing problems in how
the code has been adapted, and is also very revealing in terms of
getting up to speed with the library, and can often lead to new ideas
in terms of additional functionality or improvements.

I hope that helps to explain the package design choices better.

Best
Max

On Sat, 2 Jul 2022 at 15:01, Francesco Andreuzzi
***@***.***> wrote:
>
> Replying to the comment by @max-little under #9:
>
> Re: PyTest, we have explored this and from long experience, want to avoid
> going down the Python package interdependency route where possible. This is
> one downside of OS development e.g. lack of central control over package
> quality, and the resulting subtle and not so subtle bugs which arise as a
> result. Our preference is to ensure our code is as self-contained as
> possible. Of course, that's not always feasible.
>
> Okay, but I would strongly suggest to have testing code and testing data separated from the actual source code for the package, which is what I suggest here: #10 (comment)
>
