### 2022-07-03 max-little/GRAPL#10 @max-little
Hi Francesco

Many thanks for the observation. I don't really see this as a problem which
needs fixing, but I can agree that there are some advantages to splitting
up the tests like that. The counterpoint is that, if you split up the tests
like that, you introduce additional complexity e.g. you are making the code
harder to quickly read and understand, users will be distracted by the
myriad separate entry points to the tests.

As a compromise, I have split up the tests but then call them through a
single entry point. If users want to dig down to better localize an
assertion failure, they can now just call the individual test instead of
the main entry point.

Best
Max

On Sunday, 3 July 2022, Francesco Andreuzzi ***@***.***>
wrote:
> The scenario
> we have often encountered in writing research software like this, is
> that users tweak/adapt the code locally for their own purposes, and in
> the process tend to break it (naturally, as they are experimenting).
> Running the tests is a very quick way of diagnosing problems in how
> the code has been adapted, and is also very revealing in terms of
> getting up to speed with the library, and can often lead to new ideas
> in terms of additional functionality or improvements.
>
> pytest handles this.
>
> I see at least four problems with the current test architecture, even if
you decide to stick with this directory tree:
>
> You have a single method for all your unit tests. Users need to jump to
the code and find the line which failed in order to determine which graph
caused the assertion failure.
> Users may want to run all the tests and see what fails and what is fine,
instead of just getting the first failure (which is what happens now). I do
this very often since it helps debugging (the kind of graph which failed
may convey useful information about the type of bug).
> Debugging (i.e. using the debugger) is painful now, because it's not
possible to debug one single graph among your test graphs, without using
some conditional breakpoints or similar tricks. It's doable, but users
would need to do a lot of point-click point-click, and due to the iterative
nature of debugging this might be upsetting.
> You share local variables between different test cases. A typo like
grap_filename would be hard to spot.
>
> Code sample
>
> [...]
> grapl_filename = 'grapl/graphs/complex_dag.grapl'
> G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
> id_str, expr, isdag = algs.truncfactor(G, {'X1'}, {'X3','X4','Y1'})
> assert all([(term in expr.num) for term in
[{'Y1','C','X1'},{'X3'},{'X4'},{'C'}]])
> assert {'C','X1'} in expr.den
> assert 'C' in expr.mrg
>
> grapl_filename = 'grapl/graphs/markov1.grapl'
> G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
> assert isdag
> assert G.nd({'D'}) == {'S','I'}
> assert G.nd({'G'}) == {'D','S','I'}
> assert G.nd({'L'}) == {'D','S','I','G'}
> assert G.nd({'S'}) == {'D','I','G','L'}
> assert G.nd({'I'}) == {'D'}
> [...]
>
> More or less I get your point on the directory tree, even if I would
prefer you adapt it to the standard, (
https://packaging.python.org/en/latest/tutorials/packaging-projects).
However at least the monolithic unit test should be changed to fix the
problems I mentioned above.
>
> What I expect
>
> def test_complex_dag():
> grapl_filename = 'grapl/graphs/complex_dag.grapl'
> G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
> assert ...
> assert ...
>
> def test_markov1():
> ...
>
