### 2022-07-03 3d5c83c
Split up unit tests
Splits up unit tests in order to facilitate localization of assertion failures.
