+++ grapl/test/unit_tests.py
@@ -11,13 +11,29 @@ import grapl.algorithms as algs
 import grapl.dsl as dsl
 
 def run():
-    grapl_obj = dsl.GraplDSL()
+    test_cyclic()
+    test_topsort()
+    test_tianfactor()
+    test_fixing()
+    test_identify1()
+    test_identify2()
+    test_non_identify1()
+    test_non_identify2()
+    test_complex_dag()
+    test_markov()
 
+    print('All unit tests passed')
+
+
+def test_cyclic():
+    grapl_obj = dsl.GraplDSL()
     grapl_filename = 'grapl/graphs/cyclic_graph.grapl'
     G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
     assert not G.isdag()
     assert not G.isayclic()
 
+def test_topsort():
+    grapl_obj = dsl.GraplDSL()
     grapl_filename = 'grapl/graphs/topsort_graph.grapl'
     G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
     node_list = G.topsort()
@@ -25,6 +41,8 @@ def run():
         for parent in G.pa({node}):
             assert node_list.index(node) > node_list.index(parent)
 
+def test_tianfactor():
+    grapl_obj = dsl.GraplDSL()
     grapl_filename = 'grapl/graphs/bareinboim_2020.grapl'
     G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
     assert G.isayclic()
@@ -32,6 +50,8 @@ def run():
     assert all([(v in G.an({'D'})) for v in {'A','B','C','D'}])
     assert all([(v in G.de({'B','F'})) for v in {'B','D','E','F'}])
 
+def test_fixing():
+    grapl_obj = dsl.GraplDSL()
     grapl_filename = 'grapl/graphs/richardson_2017.grapl'
     G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
     id_expr, expr, isident = algs.idfixing(G, {'X2'}, {'X4'})
@@ -39,6 +59,8 @@ def run():
     assert G.isayclic()
     assert isident
 
+def test_identify1():
+    grapl_obj = dsl.GraplDSL()
     grapl_filename = 'grapl/graphs/shpitser_thesis1.grapl'
     G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
     id_expr, expr, isident = algs.idfixing(G, {'X'}, {'Y1','Y2'})
@@ -49,6 +71,8 @@ def run():
     assert {'X', "W1'"} in expr.den
     assert "W1'" in expr.mrg
 
+def test_identify2():
+    grapl_obj = dsl.GraplDSL()
     grapl_filename = 'grapl/graphs/shpitser_thesis2.grapl'
     G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
     id_expr, expr, isident = algs.idfixing(G, {'X'}, {'Y'})
@@ -56,6 +80,8 @@ def run():
     assert G.isayclic()
     assert isident
 
+def test_non_identify1():
+    grapl_obj = dsl.GraplDSL()
     grapl_filename = 'grapl/graphs/non_identify1.grapl'
     G = grapl_obj.readgrapl(open(grapl_filename, 'r').read())
     id_expr, expr, isident = algs.idfixing(G, {'X'}, {'Y'})
@@ -63,6 +89,8 @@ def run():
     assert G.isayclic()
     assert not isident
 
+def test_non_identify2():
+    grapl_obj = dsl.GraplDSL()
     grapl_filename = 'grapl/graphs/non_identify2.grapl'
[…]
