### 2022-07-04 844d63b
Update README.md
Small text formatting changes.
