+++ README.md
@@ -27,7 +27,7 @@ grapl_tutor.run()
 and similarly for tutorials 2-4.
 
 ## Usage example
-*Computing the front-door adjusted distribution of the causal effect of `X` on `Y` with mediator `M` (e.g. `X` -> `M` -> `Y`) with hidden/latent confounding back-door path `X` -- `Y`.*
+*Computing the front-door adjusted distribution of the causal effect of `X` on `Y` with mediator `M` (e.g. `X` &rarr; `M` &rarr; `Y`) with hidden/latent confounding back-door path `X` &mdash; `Y`.*
 
 First, create a GRAPL DSL object, then the GRAPL description of the graph in a string, and parse the string using `grapl.dsl.readgrapl` to create the graph object `G`:
 
