### 2022-07-05 max-little/GRAPL#10 @max-little
Hi Francesco

Just a quick follow up - is there anything else you wanted to raise at this
point?

Best
Max

On Mon, 4 Jul 2022 at 13:14, Francesco Andreuzzi ***@***.***>
wrote:

> Closed #10 <https://github.com/max-little/GRAPL/issues/10> as completed.
>
