### 2022-07-06 max-little/GRAPL#10 @fandreuz
Hi @max-little, as I mentioned here I started reviewing the code but then I requested some changes which should make your code more accessible to external contributors: https://github.com/openjournals/joss-reviews/issues/4534#issuecomment-1175573632
