### 2022-07-06 27df968
Expanded function descriptions in docstrings
In response to comment raised by @fAndreuzzi
