### 2022-07-06 #4534 @max-little
Hi Francesco

Thanks for the comments. Now done. Specifically, I've provided more details for the more complex algorithms/functions in the package. For the rest, the existing comments are either self-explanatory, and/or the comments within the code are sufficiently explanatory in order for the functionality to be clear. Again, this follows from literate coding principles (which do clash somewhat with the concept of docstrings).

The other consideration here which I should raise, is that this package has a specific context, that is, it is aimed at those with a background in structural causal inference, and the terminology (hence, variable/parameter naming) is geared towards that community's standards. So, the package is not intended as a substitute for the relevant textbooks (e.g. Pearl's "Causality").

Re: (_) convention, there are no functions/methods in this package which are not intended to be used.

All the best
Max
