### 2022-07-07 #4534 @max-little
Hi Francesco

Do you have any further points to raise?

Best
Max

On Tue, 5 Jul 2022 at 23:42, Francesco Andreuzzi ***@***.***>
wrote:

> Hi @max-little <https://github.com/max-little>, I started looking at your
> code and I would say that you need to invest a little bit of time in
> documentation. I'm not going to ask full docstrings on every single method,
> but I would say that there are some method where docstrings would are
> strongly needed:
>
>    - Public methods (i.e. methods that are intended to be used by the
>    user, like GraplDSL.readgrapl: Even if you feel like you would write
>    obvious things it's good practice to spend some words on explaining what's
>    going to happen and how the parameters are going to be used.
>    - Methods in algorithms.py: What are X and Y? simplifywhat? (I'm
>    referring todagfactor`). I feel that I should not need to read a paper
>    to understand the symbols you're using, so docstring here would be very
>    appreciated.
>
> Also, I would strongly suggest to prepend an underscore (_) to each
> method/function that is not intended to be used by users PEP8
> <https://peps.python.org/pep-0008/#descriptive-naming-styles>.
>
> Suggestions for docstrings from NumPy
> <https://numpydoc.readthedocs.io/en/latest/format.html>.
>
> Let me know if you need clarification on any of my points.
>
