### 2022-07-08 #4534 @max-little
Hi Francesco

Thanks for your reply, some comments below:

On Fri, 8 Jul 2022 at 00:30, Francesco Andreuzzi ***@***.***>
wrote:

> @max-little <https://github.com/max-little> I've ticked all the points
> for the paper, I'm going to review the code and try code samples in the
> next few days hopefully.
>
> For the rest, the existing comments are either self-explanatory, and/or
> the comments within the code are sufficiently explanatory in order for the
> functionality to be clear.
>
> I get your point, but this implies that your users need to look at your
> code in order to find:
>
>    1. Which features are available, besides what you show in the tutorials
>
> Well no, an IDE like VSCode automatically scans the code, clearly showing
you what features/methods are available, you get a nice list of functions
in a side panel. You can also just type help(grapl.algorithms) in an
interactive Python shell, and get a clean list of the functionality,
because minimal docstrings are provided for all features.

>
>    1. What is the meaning of the parameters, what is the expected type
>    for each parameter, etc etc
>
> But you can generally infer types from (a) the defaults, or (b) the name
of the parameter, or (c) the docstring. Detailed type-level declarations
would really clutter the code. For instance, with a function like admg.de,
typing help(grapl.admg.de) gives you:

de(self, nodes)
     |      Find the union of all descendants of a given set of nodes, and
the nodes themselves.

So it's entirely self-explanatory: the function admg.de requires a node
parameter, which is a set of nodes. It's really as simple as that.

>
>
> I guess it's rare to find users willing (or happy) to jump into half an
> hour of diving into the source code to get to use an open source package.
> For me, I would definitely, but I try not to assume this when I write open
> source software.
>
As I said, I think you are missing the context. This is research-level
software, it isn't intended to be used by those who have no background at
all in the topic. As such, providing extensive docstrings with content down
to the type declaration level, just obscures the code making it harder to
understand. The amount of comments would dwarf the actual code in size.

> Again, this follows from literate coding principles (which do clash
> somewhat with the concept of docstrings).
>
> Docstrings integrate nicely with the Python environment (e.g. IDE) and
> most importantly enable using tools like Sphinx
> <https://www.sphinx-doc.org/en/master/> which auto-generate documentation
> pages in HTML format. For me the initial effort you require to your users
> is quite a lot, but I'm going to wait for the opinion of @Athene-ai
> <https://github.com/Athene-ai> on this topic.
>
That's not really the point: literate coding is about creating
"self-documenting" code, to the extent that is possible. I agree that
integrating with tools like IDEs and document generators such as Sphinx,
provides a nice level of convenience with "referencing automation", but
this can't be a substitute for understanding the relevant concepts. That
would be a different kind of package.

Once again, thanks for your comments/contributions, it's helpful to have
these discussions.

Best
Max

--
Max Little (www.maxlittle.net)
Associate Professor, University of Birmingham
Turing Fellow, Alan Turing Institute
TED Fellow (fellows.ted.com/profiles/max-little)
Visiting Associate Professor, MIT
Author: Machine Learning for Signal Processing, Oxford University Press
global.oup.com/academic/product/machine-learning-for-signal-processing-9780198714934
Room 138, School of Computer Science
University of Birmingham
Birmingham B15 2TT
UK +44 7710 609564
Skype: dr.max.little
