### 2022-07-08 #4534 @max-little
Hi Francesco

> As I said, I think you are missing the context. This is research-level
> software, it isn't intended to be used by those who have no background at
> all in the topic.
>
> I expect some users would be blocked by lack of understanding in Python programming rather than in the topic GRAPL addresses. Research-level software means that if I'm a C++ programmer it should not take me too much time to start using what you provide. Or at least this is what I would expect from a JOSS publication.

I don't get that, to be honest. I think that, if the code is
well-structured, you shouldn't have that much difficulty in going from
e.g. Python to C++ as they are both imperative languages, albeit with
their own idiosyncrasies, although I agree you would struggle if it
were written in Haskell and you had no background in functional
programming at all. I think if you are asking code to be written such
that someone who has no knowledge of the language at all of the
language in which it is written, I don't really see how they can use
it without getting lost anyway. In any case, I'm just going on JOSS's
own description:

"JOSS publishes articles about research software. This definition
includes software that: solves complex modeling problems in a
scientific context (physics, mathematics, biology, medicine, social
science, neuroscience, engineering); supports the functioning of
research instruments or the execution of research experiments;
extracts knowledge from large data sets; offers a mathematical
library; or similar.

JOSS submissions must:
* Be open source (i.e., have an OSI-approved license).
* Have an obvious research application.
* Be feature-complete (no half-baked solutions) and be designed for
maintainable extension (not one-off modifications)."

It doesn't mention anything about how much time it takes a user to
pick up the software, or the level of expected expertise in the topic,
and I assume this is for good reason because some topics are more
complex and require extensive background knowledge, so the package
would, presumably require a lot more time to pick up. Others are just
more generic because they are less specialist (see below for more on
this).

> As such, providing extensive docstrings with content down
> to the type declaration level, just obscures the code making it harder to
> understand.
>
> I would suggest taking a look at some other open source projects published in order to see if the kind of documentation you provide is similar to what other authors are doing. I took some links from the first 4 papers I found on the main page of JOSS:
>
> https://lanl.github.io/spiner/main/index.html
> https://rbturnbull.github.io/ausdex/
> https://python-kmedoids.readthedocs.io/en/latest/
> https://github.com/nasa/simupy-flight/blob/master/simupy_flight/__init__.py
>
> As you can see they all provide extensive explanations and examples, moreover in the Module index detailed descriptions (with parameters and return value) are provided for a big subset of the functions provided by the package. The last one is the package which provides the least complete documentation, but as you can see in the file I linked for you the author provided very detailed docstrings for all the public functions, to be used directly by the user.
>
> Clearly my survey is not complete and depending on the reviewer you may find different levels of documentation accepted for JOSS publications. This is why I said I'd like to wait for my fellow reviewer before asking additional improvements on documentation.

I did already take quite a long look at other repositories before
preparing the code for JOSS. Your informal survey has picked out much
more "generic" software packages. By contrast a package which is much
more "domain specific"/specialist topic, such as this one:
https://gitlab.com/thartwig/asloth

has a lot _less_ comment detail than GRAPL, e.g.:
https://gitlab.com/thartwig/asloth/-/blob/main/scripts/plot_asloth.py
https://gitlab.com/thartwig/asloth/-/blob/main/src/SF_routines.F90

I don't know the topic of that package myself, but there must be a
reason why the code is not cluttered with extraneous comments; I would
suggest that is because, like GRAPL, this is intended for specialists
who can read the code and understand what the software is doing
because it is, effectively presented in a "language" they understand
already.

Hope that helps explain what I mean more clearly.

All the best
Max
