### 2022-07-08 #4534 @max-little
Another relevant example package would be this one:
https://gitlab.ruhr-uni-bochum.de/reichp2y/proppy

No docstrings at all on some of the major entry points (e.g. see simulation.run_simulation):
https://gitlab.ruhr-uni-bochum.de/reichp2y/proppy/-/blob/master/proppy/simulation.py
https://gitlab.ruhr-uni-bochum.de/reichp2y/proppy/-/blob/master/proppy/source.py

There is additional documentation, but it's generated automatically from the code, so it's no more detailed than the docstrings provided.

Again, I presume this is because, the code is largely self-explanatory.

Best
Max
