### 2022-07-08 #4534 @Athene-ai
Hi all,

I think that it is quite tricky to understand how to move from Phyton to
C++.

Thus could be better explained
