### 2022-07-08 #4534 @max-little
OK, thanks both.
So, it looks like the consensus is you would be satisfied with: expanded docstrings for *most* methods/functions?
Best,
Max
