+++ grapl/admg.py
@@ -11,7 +11,11 @@ from queue import Empty
 import grapl.util as util
 
 class ADMGNode():
-    """An ADMG node (random variable) object."""
+    """An ADMG node (random variable) object. The member variables, parents, children,
+       bidirects are Sets of Strings, each string is the name of a random variable (node)
+       in the graph. Bidirects are (nodes) connected to this one via a latent edge. The
+       properties member is an optional Set of properties.
+    """
     def __init__(self, parents={}, bidirects={}, children={}, properties={}):
         self.parents = set(parents)
         self.children = set(children)
@@ -19,24 +23,34 @@ class ADMGNode():
         self.properties = set(properties)
 
 class ADMG():
-    """An ADMG graph object."""
+    """An ADMG graph object. The title member variable (String) is the name of graph
+       (mostly for display purposes). The vars member is a Dict of String-ADMGNode pairs,
+       representing his graph's random variables (ADMGNodes), indexed by node name (String).
+    """
     def __init__(self, title=''):
         self.title = title
         self.vars = dict()
 
     def settitle(self, title=''):
-        """Set graph title."""
+        """Set graph title (String)."""
         self.title = title
 
     def addvar(self, name='', parents={}, bidirects={}, children={}, properties={}):
-        """Add a variable to a graph with optional parents, bidirects and/or children."""
+        """Add a new random variable to the graph with string name, with optional parents, bidirects
+           children and/or properties. The name must not be empty.
+           Each random variable (parents, bidirects, children) is a Set of Strings,
+           each string representing a random variable in the graph.
+        """
         if name != '':
             self.vars[name] = ADMGNode(parents,bidirects,children,properties)
         else:
             raise ValueError('admg:addvar: Invalid node name')
 
     def addedges(self, name='', parents={}, bidirects={}, children={}):
-        """Add parents, bidirects and/or child edges to a graph."""
+        """Add (optional) new edges connecting parents, bidirects and/or child random variables,
+           to the random variable given in name (which must not be empty).
+           Edges are Sets of Strings, each string representing a random variable in the graph.
+        """
         if name != '':
             self.vars[name].parents = self.vars[name].parents.union(parents)
             self.vars[name].children = self.vars[name].children.union(children)
@@ -45,7 +59,12 @@ class ADMG():
             raise ValueError('admg:addedges: Invalid node name')
 
     def connect(self):
-        """Ensure all parent/child/bidirect edges are consistent across the graph."""
+        """Ensure all parent/child/bidirect edges are consistent across the graph.
+           Explicitly: every parent of a random variable, must have that variable
+           as a child, and every random variable connected by a latent (bidrected)
+           edge to another random variable, must have a bidirected edge to that
+           first random variable.
+        """
         for node in set(self.vars):
             for p in self.pa({node}):
                 self.vars[p].children = self.ch({p}).union({node})
@@ -53,7 +72,7 @@ class ADMG():
                 self.vars[b].bidirects = self.bi({b}).union({node})
 
     def display(self):
-        """Pretty-print a text representation of the graph."""
+        """Pretty-print a text representation of the graph (ADMG) object."""
         nodes = set(self.vars)
         print('Title: ' + self.title)
         print('Vars: ' + util.csepstr(nodes))
@@ -74,11 +93,17 @@ class ADMG():
                 print(node,'<-->', util.csepstr(bidirects))
 
     def nodes(self):
-        """Return set of all nodes (variables)."""
+        """Return set of all random variables (nodes) in the graph.
+           This is represented as a Set of Strings, each string being the name of
+           a random variable.
+        """
         return set(self.vars)
 
     def sub(self, subvars):
-        """Find a subgraph of a graph."""
+        """Find a subgraph of a graph. The subvars parameter is a Set of Strings,
+           each string naming a random variable. If there are no such random variables
[…]
