### 2022-07-12 4f796b6
Documentation and related updates.
Extensive updates to docstrings including types of input and output parameters. Labelling of 'internal' functions with '_'.
