### 2022-07-12 f1c2347
Documentation and related updates.
Extensive updates to docstrings including types of input and output parameters. Labelling of 'internal' functions with '_'.
