### 2022-07-12 30776fa
Documentation and related updates.
Extensive updates to docstrings including types of input and output parameters. Labelling of 'internal' functions with '_'.
