### 2022-07-12 #4534 @max-little
Hi @fAndreuzzi @Athene-ai (cc @Nikoleta-v3)
As requested, I have now made substantial changes as follows:
1. Extensive updates to docstrings.
2. Methods, member variables and classes all now fully documented.
3. Parameter and return types detailed.
4. Further detail on functionality for all methods.
5. As per our discussion, on further inspection, there *are* some internal methods which would be best described as "hidden" i.e. should not be called by the user e.g. as closures (in contradiction to what I was claiming earlier). Therefore, as requested, these methods are labelled with '_' and do not have docstrings (in contrast to the "public" methods), following Python conventions.
I think the code is now a lot more clearly presented and described, and should be easier to maintain.
Hope this is satisfactory.
Best
Max
