### 2022-07-12 56740fd
Update algorithms.py
Docstring details (parameter and return types)
