### 2022-07-22 d706a1c
easier tutorial usage. code cleanup. improve naming
