### 2022-07-22 #4534 @fAndreuzzi
Hi @max-little, first of all thank you for addressing the issues mentioned by @Athene-ai and me. I ticked some more checkboxes in my review.

Looking at the tutorials, I have some more observations:
- I don't feel the need to import and then run tutorials in two separate commands, I fixed this in max-little/GRAPL#11
- In my opinion I should not need to open the code for a tutorial (or run it) to see what that tutorial is about. I suggest adding a table in the section **Installation and getting started** where for each tutorial you explain very briefly what you do in that tutorial

Also, I tried to run this snippet:
```python
import grapl.algorithms as algs
import grapl.dsl as dsl

grapl_obj = dsl.GraplDSL()
G = grapl_obj.readgrapl('hello')
```

and received this error in response:
```python
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/francescoandreuzzi/Programming/python/GRAPL/grapl/dsl.py", line 111, in readgrapl
    self.parser.parse(scm_string)
  File "/Users/francescoandreuzzi/Programming/python/GRAPL/grapl/ply/yacc.py", line 333, in parse
    return self.parseopt_notrack(input, lexer, debug, tracking, tokenfunc)
  File "/Users/francescoandreuzzi/Programming/python/GRAPL/grapl/ply/yacc.py", line 1201, in parseopt_notrack
    tok = call_errorfunc(self.errorfunc, errtoken, self)
  File "/Users/francescoandreuzzi/Programming/python/GRAPL/grapl/ply/yacc.py", line 192, in call_errorfunc
    r = errorfunc(token)
  File "/Users/francescoandreuzzi/Programming/python/GRAPL/grapl/dsl.py", line 102, in p_error
    raise SyntaxError("Grapl syntax error at line: " + str(p.lineno) + ", near: " + p.value)
AttributeError: 'NoneType' object has no attribute 'lineno'
```

I get that `hello` is definitely not the kind of output that you expect, but I feel like you should catch this mistake from the user before it triggers an exception and inform the user that (maybe) the format they used for the input was not appropriate.

Let me know what you think about this
