### 2022-07-25 #4534 @max-little
Thanks @fAndreuzzi . I've merged in your suggested changes to the tutorials and the readme, which look sensible to me. I'll see about adding a short explanatory table for the tutorials.

Re: syntax error, well, obviously, if you just put random text into the DSL parser without understanding the language at all, then you will likely just get lexical, syntax or semantic errors, which would happen with any such domain-specific language (DSL). The DSL code does, however, include complete error handling. This is all pretty standard for any DSL. Unfortunately, in this case you discovered a bug whereby the case of missing input was generating a spurious exception. This is now fixed in the latest version.

To explain in more detail, in your test the string 'hello' on its own doesn't have an end of command (EOC) terminator. In fact, 'hello;' (with the semicolon command terminator) would be a valid GRAPL DSL string, as you can check by running:

`import grapl.dsl as dsl

grapl_obj = dsl.GraplDSL()
G = grapl_obj.readgrapl('hello;')`

(What you will get is a graph G with a single node called 'hello', you can infer this from the tutorials and/or see precisely why by looking at the EBNF for the DSL, in dsl.py).

In any case, what you will now get, if you don't terminate a command, is the following:

`grapl_obj.readgrapl('hello')
Grapl syntax error: unexpected end of input`

In general, GRAPL error handling reports the line and some characters near where the error occurred (again standard for this kind of DSL). Furthermore, you can always debug a string using `dsl.displaygrapltokens`, reporting the output of the lexical parsing, which can help to narrow down bugs.
