### 2022-07-26 ffb428c
Update README.md
More detailed description of content of tutorials.
