### 2022-07-27 #4534 @fAndreuzzi
Hey @max-little,

> To explain in more detail, in your test the string 'hello' on its own doesn't have an end of command (EOC) terminator. In fact, 'hello;' (with the semicolon command terminator) would be a valid GRAPL DSL string, as you can check by running [...]

Thank you very much for the explanation, now it's clearer to me.

> In any case, what you will now get, if you don't terminate a command, is the following:
>
> ```python
> grapl_obj.readgrapl('hello')
> Grapl syntax error: unexpected end of input
> ```

Perfect! Thanks for the change.

I've just opened max-little/GRAPL#12 for a minor improvement in the readability of the README file (just to let the user quickly recognize where are the tutorials, how to run them and what's the content). Feel free to reject or merge depending on what you need.

My review is over I think, all my checkpoints are ticked @Nikoleta-v3
