### 2022-08-01 #4534 @max-little
Hi @Athene-ai (cc @Nikoleta-v3 )

Just a quick follow-up to ask if there were any additional issues you wanted to raise, or are you happy with the now revised codebase?

Best
Max
