### 2022-08-03 #4534 @Nikoleta-v3
@max-little what do you think about this:

> I see that your code is not on PyPI, do you plan to upload it there such that your users can install GRAPL via pip install grapl?
