### 2022-08-03 #4534 @max-little
@Nikoleta-v3 OK there's now also the option to install through PyPI:
https://pypi.org/project/grapl-causal/1.4/
