### 2022-08-03 #4534 @Nikoleta-v3
Nothing else you can do for now 😃 I pinged the editorial board. One of the editors will check that everything is okay with the paper, and then publish it. They might have a few more comments but at this point they will be very minor.
