### 2022-10-11 #4534 @max-little
> Yes, most meaning what is going to be used by the user. You should document also classes, as they support docstrings too.

OK I'll do that, it's basically going to be all methods/functions and classes.

> I would also suggest to split functions into smaller, more maintainable ones, it looks quite weird and dangerous to me that you don't have functions which are not going to be used by the user. Even if you do not share any code among different functions, you should try to pack your logic into smaller, single-responsibility functions.

As I've said, there are no functions/methods which should not be
accessed by the user. This is the logic of structural causal modelling
and inference. Can you perhaps explain to me which parts of the
functionality of the software shouldn't be accessed by the user, and
give a convincing reason why, in terms of the logic of structural
causal modelling and inference?

Best
Max
